Create conftest.py for common fixtures
